Move cluster API to root
