+++ src/fairseq2/cli/utils/cluster.py
@@ -8,8 +8,8 @@ from __future__ import annotations
 
 import os
 
+from fairseq2.cluster import ClusterHandler, ClusterResolver
 from fairseq2.context import RuntimeContext
-from fairseq2.recipes.cluster import ClusterHandler, ClusterResolver
 
 
 def set_torch_distributed_variables(context: RuntimeContext, cluster: str) -> None:
