+++ src/fairseq2/setup/_lib.py
@@ -16,13 +16,13 @@ from fairseq2.assets import (
 )
 from fairseq2.assets.setup import register_assets
 from fairseq2.chatbots.setup import register_chatbots
+from fairseq2.cluster import register_clusters
 from fairseq2.context import RuntimeContext, set_runtime_context
 from fairseq2.datasets.setup import register_datasets
 from fairseq2.extensions import run_extensions
 from fairseq2.metrics import register_metric_descriptors
 from fairseq2.metrics.recorders import register_metric_recorders
 from fairseq2.profilers import register_profilers
-from fairseq2.recipes.cluster import register_clusters
 from fairseq2.recipes.setup import register_recipes
 from fairseq2.setup._generation import (
     _register_beam_search_algorithms,
